+++ docs/conf.py
@@ -28,7 +28,10 @@ sys.path.append(os.path.dirname(os.path.abspath('.')))
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx']
+
+# link to sqlalchemy docs
+intersphinx_mapping = {'http://www.sqlalchemy.org/docs/05/': None}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']