+++ migrate/changeset/__init__.py
@@ -4,5 +4,11 @@
 
    .. [#] SQL Data Definition Language
 """
+import sqlalchemy
+
 from migrate.changeset.schema import *
 from migrate.changeset.constraint import *
+
+sqlalchemy.schema.Table.__bases__ += (ChangesetTable, )
+sqlalchemy.schema.Column.__bases__ += (ChangesetColumn, )
+sqlalchemy.schema.Index.__bases__ += (ChangesetIndex, )